Move NotConnectedRestriction to the global components folder
